Week 3 Layout/Styling HW
This is really late, i’m sorry!!! The jquery additions/HW for this week
will be coming in a bit, I’m keeping submits separate to keep things a
little more organized. I feel pretty confident about most things here,
layout is becoming much more fluent for me. I had a little trouble
laying out the bottoms of the paragraphs with the borders so there
would be enough room in the design. Also, the pseudo class of
.firstchild was not functioning for me but .firstcharacter did the
trick.